refactored media helpers
